requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argilla==1.4.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
black==21.4b2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cohere==4.0.1
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
dataclasses-json==0.5.7
Deprecated==1.2.13
detectron2==0.6
effdet==0.3.0
et-xmlfile==1.1.0
fastapi==0.94.1
filelock==3.9.1
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.39.0
frozenlist==1.3.3
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.16.2
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.2
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.107
layoutparser==0.3.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.2
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
portalocker==2.7.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.6
pydot==1.4.2
Pygments==2.14.0
pypandoc==1.11
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
qdrant-client==1.0.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
timm==0.6.12
tokenizers==0.13.2
toml==0.10.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1
typing-extensions==4.5.0
typing-inspect==0.8.0
unstructured==0.5.4
unstructured-inference==0.2.11
urllib3==1.26.15
uvicorn==0.21.0
Wand==0.6.11
Werkzeug==2.2.3
wrapt==1.14.1
XlsxWriter==3.0.9
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0
